+++ transaction-status/src/parse_token/extension/confidential_transfer.rs
@@ -1,9 +1,9 @@
 use {
     super::*,
+    solana_zk_sdk_pod::encryption::elgamal::PodElGamalPubkey,
     spl_token_2022_interface::{
         extension::confidential_transfer::instruction::*,
         instruction::{decode_instruction_data, decode_instruction_type},
-        solana_zk_sdk::encryption::pod::elgamal::PodElGamalPubkey,
     },
 };
 
@@ -598,18 +598,16 @@ mod test {
         solana_instruction::{AccountMeta, Instruction},
         solana_message::Message,
         solana_pubkey::Pubkey,
+        solana_zk_sdk_pod::encryption::{
+            auth_encryption::PodAeCiphertext, elgamal::PodElGamalCiphertext,
+        },
         spl_token_2022_interface::{
             extension::confidential_transfer::instruction::{
                 initialize_mint, inner_configure_account, inner_empty_account, update_mint,
             },
-            solana_zk_sdk::{
-                encryption::pod::{
-                    auth_encryption::PodAeCiphertext, elgamal::PodElGamalCiphertext,
-                },
-                zk_elgamal_proof_program::proof_data::{
-                    BatchedGroupedCiphertext3HandlesValidityProofData, BatchedRangeProofU128Data,
-                    CiphertextCommitmentEqualityProofData, ZeroCiphertextProofData,
-                },
+            solana_zk_elgamal_proof_interface::proof_data::{
+                BatchedGroupedCiphertext3HandlesValidityProofData, BatchedRangeProofU128Data,
+                CiphertextCommitmentEqualityProofData, ZeroCiphertextProofData,
             },
         },
         spl_token_confidential_transfer_proof_extraction::instruction::ProofLocation,